substracted the controllers from the server into it's own folder
