+++ controllers/chits_controller.rb
@@ -1,0 +1,6 @@
+post '/chits' do
+  username = params["username"]
+  text = params["text"]
+  Chit.create(:username => username, :text => text)
+  redirect to('/')
+end
